requirements.txt
elevenlabs
flask[async]
Flask_Cors
Flask_SocketIO
openai
pandas
python-dotenv
scikit-learn
sentence_transformers